fix: detail
