+++ packages/yike-design-ui/src/components/timeline/index.ts
@@ -1,5 +1,5 @@
 import Timeline from './src/timeline.vue';
-import TimelineItem from './src/timelineitem.vue';
+import TimelineItem from './src/timeline-item.vue';
 import { withInstall } from '../utils/index';
 
 export const YkTimeline = withInstall(Timeline);
