+++ packages/yike-design-ui/src/components/timeline/src/internal.ts
@@ -1,0 +1,7 @@
+import { InjectionKey } from 'vue';
+import { TimelinePlacement } from '../../utils';
+
+export const timelineContextKey: InjectionKey<{
+  horizontal: boolean;
+  placement: TimelinePlacement;
+}> = Symbol('scrollbarContext');
