--- packages/yike-design-ui/src/components/timeline/src/timelineitem.ts
@@ -1,9 +1,0 @@
-import { TimelineType } from '../../utils/constant';
-export type TimelineItemProps = {
-  time?: string;
-  title?: string;
-  content?: string;
-  lineType?: TimelineType;
-  dotColor?: string;
-  pending?: boolean;
-};
